# component pkg

+++ global/sunny-framework-js/main.sh
@@ -5,6 +5,10 @@ registry=http://nexus.czy21-internal.com/repository/npm-hosted/
 
 
 for pkg in `ls -d packages/*`;do
+  if [ "$(basename $pkg)" != "vue" ];then
+    continue
+  fi
+
   rm -rf $pkg/dist
   pkg_name=$(basename $pkg)
   npm run build -w @sunny-framework-js/$pkg_name
